refactor | todo item 분리
