+++ src/interfaces/ITodoData.ts
@@ -1,0 +1,6 @@
+export interface ITodoData {
+  id: number
+  todo: string
+  isCompleted: boolean
+  userId: number
+}
